django/mysite/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Deprecated==1.2.13
distlib==0.3.2
Django==3.2.8
django-cors-headers==3.10.0
django-dramatiq==0.10.0
djangorestframework==3.12.4
dramatiq==1.12.0
Envelopes==0.4
filelock==3.0.12
Flask==2.0.2
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mpmath==1.2.1
numpy==1.21.2
Pillow==8.4.0
platformdirs==2.3.0
prometheus-client==0.12.0
python-constraint==1.4.0
pytz==2021.3
raven==6.10.0
redis==4.0.2
requests==2.25.1
scipy==1.7.1
six==1.16.0
sqlparse==0.4.2
sympy==1.8
urllib3==1.26.6
virtualenv==20.7.2
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.13.3
